Fish/requirements.txt
aiofiles==24.1.0
alembic==1.16.4
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pefile==2023.2.7
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.10.1
pydantic_core==2.33.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5.0
PyMySQL==1.1.1
PyPDF2==3.0.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-docx==1.1.2
python-dotenv==1.1.1
python-jose==3.4.0
python-multipart==0.0.20
pywin32-ctypes==0.2.3
requests==2.32.4
rsa==4.9
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tk==0.1.0
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.5.0
uvicorn==0.34.0
